refactor: split admin menu handlers into categories and items
